+++ drivers/spi/spi_nxp_lpspi/Kconfig
@@ -16,6 +16,7 @@ if !SPI_RTIO
 config SPI_MCUX_LPSPI_DMA
 	bool "MCUX LPSPI SPI DMA Support"
 	select DMA
+	depends on $(dt_compat_any_has_prop,$(DT_COMPAT_NXP_LPSPI),dmas)
 	help
 	  Enable the SPI DMA mode for SPI instances
 	  that enable dma channels in their device tree node.
@@ -23,6 +24,7 @@ config SPI_MCUX_LPSPI_DMA
 config SPI_MCUX_LPSPI_NORMAL
 	bool "NXP MCUX LPSPI driver"
 	default y
+	depends on $(dt_compat_any_not_has_prop,$(DT_COMPAT_NXP_LPSPI),dmas) || !SPI_MCUX_LPSPI_DMA
 	help
 	  Use the traditional (non-RTIO) SPI driver for NXP LPSPI.
 endif # !SPI_RTIO
